chore(refactor): make Filters function based component
