+++ src/Components/SmartComponents/CVEs/VulnerabilitiesTableToolbar.js
@@ -10,21 +10,20 @@ import {
     kebabItemEditBusinessRisk
 } from '../../PresentationalComponents/DownloadReportKebab/KebabItems';
 import PaginationWrapper from '../../PresentationalComponents/PaginationWrapper/PaginationWrapper';
-import Filters from '../Filters/Filters';
+import Filters from '../../PresentationalComponents/Filters/Filters';
 import Remediation from '../Remediation/Remediation';
 import SelectAllCheckbox from '../SelectAllCheckbox/SelectAllCheckbox';
 import { CVETableContext } from './VulnerabilitiesCves';
+import { filtersCVSSScore, filtersPublishDate, filtersSeverity, filtersShowAll } from '../../../Helpers/constants';
 
 class VulnerabilitiesToolbarWithContext extends Component {
     static propTypes = {
-        showAllCheckbox: propTypes.bool,
         showRemediationButton: propTypes.bool,
         entity: propTypes.object,
         context: propTypes.object
     };
 
     static defaultProps = {
-        showAllCheckbox: false,
         showRemediationButton: false,
         totalNumber: 0,
         apply: () => undefined,
@@ -32,7 +31,7 @@ class VulnerabilitiesToolbarWithContext extends Component {
     };
 
     render() {
-        const { showAllCheckbox, showRemediationButton, entity, context } = this.props;
+        const { showRemediationButton, entity, context } = this.props;
         const { cves, params, methods } = context;
         const { selectedCves } = params;
         const selectedCvesCount = selectedCves && selectedCves.length;
@@ -55,10 +54,15 @@ class VulnerabilitiesToolbarWithContext extends Component {
                         fetchResource={ops => methods.fetchResource({ ...params, ...ops, show_all: !params.show_all })}
                     />
                     <Filters
+                        searchPlaceholder="Find a CVE…"
                         apply={methods.apply}
-                        showStatusList={Boolean(entity)}
-                        showAllCheckbox={showAllCheckbox}
                         filterValues={params}
+                        filterCategories={[
+                            filtersShowAll,
+                            filtersCVSSScore,
+                            filtersSeverity,
+                            filtersPublishDate
+                        ]}
                     />
                     {showRemediationButton && <Remediation systemId={entity.id} selectedCves={selectedCves} />}
                     <BaseKebab dropdownItems={kebabOptions} />
