+++ src/Components/SmartComponents/SystemCves/SystemCveTableToolbar.js
@@ -10,21 +10,20 @@ import {
     kebabItemEditBusinessRisk
 } from '../../PresentationalComponents/DownloadReportKebab/KebabItems';
 import PaginationWrapper from '../../PresentationalComponents/PaginationWrapper/PaginationWrapper';
-import Filters from '../Filters/Filters';
+import Filters from '../../PresentationalComponents/Filters/Filters';
 import Remediation from '../Remediation/Remediation';
 import SelectAllCheckbox from '../SelectAllCheckbox/SelectAllCheckbox';
 import { CVETableContext } from './SystemCves';
+import { filtersCVSSScore, filtersPublishDate, filtersSeverity, filtersStatus } from '../../../Helpers/constants';
 
 class SystemCveToolbarWithContext extends Component {
     static propTypes = {
-        showAllCheckbox: propTypes.bool,
         showRemediationButton: propTypes.bool,
         entity: propTypes.object,
         context: propTypes.object
     };
 
     static defaultProps = {
-        showAllCheckbox: false,
         showRemediationButton: false,
         totalNumber: 0,
         apply: () => undefined,
@@ -32,7 +31,7 @@ class SystemCveToolbarWithContext extends Component {
     };
 
     render() {
-        const { showAllCheckbox, showRemediationButton, entity, context } = this.props;
+        const { showRemediationButton, entity, context } = this.props;
         const { cves, params, methods } = context;
         const { selectedCves } = params;
         const selectedCvesCount =
@@ -58,10 +57,15 @@ class SystemCveToolbarWithContext extends Component {
                         />
                     )}
                     <Filters
+                        searchPlaceholder="Find a CVE…"
                         apply={methods.apply}
-                        showStatusList={Boolean(entity)}
-                        showAllCheckbox={showAllCheckbox}
                         filterValues={params}
+                        filterCategories={[
+                            filtersCVSSScore,
+                            filtersSeverity,
+                            filtersPublishDate,
+                            filtersStatus
+                        ]}
                     />
                     {showRemediationButton && <Remediation systemId={entity.id} selectedCves={selectedCves} />}
                     <BaseKebab dropdownItems={kebabOptions} />
